+++ source/mplayer/osdep/timer-darwin.c
@@ -33,8 +33,6 @@ static double timebase_ratio;
 
 const char timer_name[] = "Darwin accurate";
 
-
-
 /* the core sleep function, uses floats and is used in MPlayer G2 */
 static float sleep_accurate(float time_frame)
 {
@@ -48,70 +46,69 @@ static float sleep_accurate(float time_frame)
 /* wrapper for MPlayer G1 */
 int usec_sleep(int usec_delay)
 {
-  return sleep_accurate(usec_delay / 1e6) * 1e6;
+	return sleep_accurate(usec_delay / 1e6) * 1e6;
 }
-
 
 /* current time in microseconds */
 unsigned int GetTimer(void)
 {
-  return (unsigned int)(uint64_t)(mach_absolute_time() * timebase_ratio * 1e6);
+	return (uint64_t)(mach_absolute_time() * timebase_ratio * 1e6);
 }
 
 /* current time in milliseconds */
 unsigned int GetTimerMS(void)
 {
-  return (unsigned int)(uint64_t)(mach_absolute_time() * timebase_ratio * 1e3);
+	return (uint64_t)(mach_absolute_time() * timebase_ratio * 1e3);
 }
 
 /* time spent between now and last call in seconds */
 float GetRelativeTime(void)
 {
-  double last_time = relative_time;
+	double last_time = relative_time;
 
-  if (!relative_time)
-    InitTimer();
+	if (!relative_time)
+		InitTimer();
 
-  relative_time = mach_absolute_time() * timebase_ratio;
+	relative_time = mach_absolute_time() * timebase_ratio;
 
-  return (float)(relative_time-last_time);
+	return relative_time - last_time;
 }
 
 /* initialize timer, must be called at least once at start */
 void InitTimer(void)
 {
-  struct mach_timebase_info timebase;
+	struct mach_timebase_info timebase;
 
-  mach_timebase_info(&timebase);
-  timebase_ratio = (double)timebase.numer / (double)timebase.denom
-    * (double)1e-9;
+	mach_timebase_info(&timebase);
+	timebase_ratio = (double)timebase.numer / (double)timebase.denom
+		* 1e-9;
 
-  relative_time = (double)(mach_absolute_time() * timebase_ratio);
+	relative_time = (double)(mach_absolute_time() * timebase_ratio);
 }
 
 #if 0
 #include <stdio.h>
 
 int main(void) {
-  int i,j, r, c = 200;
-  long long t = 0;
+	int i, j, r, c = 200;
+	long long t = 0;
 
-  InitTimer();
+	InitTimer();
 
-  for (i = 0; i < c; i++) {
-    const int delay = rand() / (RAND_MAX / 1e5);
-    j = GetTimer();
+	for (i = 0; i < c; i++) {
+		const int delay = rand() / (RAND_MAX / 1e5);
+		j = GetTimer();
 #if 1
-    r = usec_sleep(delay);
+		r = usec_sleep(delay);
 #else
-    r = sleep_accurate(delay / 1e6) * 1e6;
+		r = sleep_accurate(delay / 1e6) * 1e6;
 #endif
-    j = (GetTimer() - j) - delay;
-    printf("sleep time:%8i %5i (%i)\n", delay, j, j - r);
-    t += j - r;
-  }
-  fprintf(stderr, "average error:\t%lli\n", t / c);
+		j = (GetTimer() - j) - delay;
+		printf("sleep time:%8i %5i (%i)\n", delay, j, j - r);
+		t += j - r;
+	}
+	fprintf(stderr, "average error:\t%lli\n", t / c);
 
-  return 0;
+	return 0;
 }
 #endif
